source_code/dashboard_gui/requirements.txt
eel==0.16.0
paho_mqtt==1.6.1
python-dotenv==1.0.1
